TST/CLN: deduplicate fixture from test_to_latex.py (#26603)
